% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Shows a map of a given state with a dot for every accident that happened in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state (integer) to plot the data for}

\item{year}{The year to map the data for (integer)}
}
\value{
None
Shows a message when an invalid state number is given or when no accidents were found to plot
}
\description{
Shows a map of a given state with a dot for every accident that happened in a given year
}
\examples{
\dontrun{
  fars_map_state(1, 2014)
}

}
