% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from a CSV file into a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(string) Filename of the file to read}
}
\value{
The data from the given CSV file as a data frame tbl (see tibble-package)
Stops when the input file does not exist.
}
\description{
Reads data from a CSV file into a data frame tbl
}
\examples{
\dontrun{
  fars_read('accident_2013.csv.bz2')
}

}
