% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data and retrieves the month and year of each accident from a given list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years (integers) to get the data for}
}
\value{
A list with a table (data frame) per year with the columns MONTH and year of each accident
Shows a warning when an invalid year is specified or a year for which no data is available
}
\description{
Reads FARS data and retrieves the month and year of each accident from a given list of years
}
\examples{
\dontrun{
  fars_read_years(2013:2014)
}

}
