% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads FARS data and retrieves the number of accidents per month for each wanted year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years (integers) to get the data for}
}
\value{
A table (data frame) with, for each asked year, the number of accidents per month
Shows a warning when an invalid year is specified or a year for which no data is available
}
\description{
Reads FARS data and retrieves the number of accidents per month for each wanted year
}
\examples{
\dontrun{
  fars_summarize_years(2013:2014)
}

}
