% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the filename to load FARS data of a particular year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year (integer) to get the filename for}
}
\value{
The filename (string) of the FARS data for a particular year
}
\description{
Creates the filename to load FARS data of a particular year
}
\examples{
\dontrun{
  make_filename(2014)
}

}
